requirements.txt
alembic==1.9.1
anyio==3.6.2
aoc-lube==0.2.6
apturl==0.5.2
async-generator==1.10
attrs==22.1.0
bcrypt==3.1.7
beautifulsoup4==4.11.1
blinker==1.4
Brlapi==0.7.0
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
Click==7.0
colorama==0.4.3
cryptography==38.0.4
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==2.2.1
duplicity==0.8.12.0
ecdsa==0.18.0
email-validator==1.3.0
entrypoints==0.3
exceptiongroup==1.0.4
fastapi==0.88.0
fasteners==0.14.1
Flask==1.1.1
future==0.18.2
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.2
httplib2==0.14.0
httptools==0.5.0
httpx==0.23.1
idna==2.8
indicator-cpufreq==0.2.2
itsdangerous==1.1.0
Jinja2==3.1.2
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.1
monotonic==1.5
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
orjson==3.8.2
outcome==1.2.0
packaging==21.3
paramiko==2.6.0
passlib==1.7.4
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
pydantic==1.10.2
pygame==2.1.2
PyGObject==3.36.0
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.28.1
requests-unixsocket==0.2.0
rfc3986==1.5.0
rsa==4.9
SecretStorage==2.3.1
selenium==4.7.2
simplejson==3.16.0
sip==4.19.21
six==1.14.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.44
starlette==0.22.0
systemd-python==234
tomlkit==0.11.6
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
ubuntu-advantage-tools==27.12
ubuntu-drivers-common==0.0.0
ufw==0.36
ujson==5.6.0
unattended-upgrades==0.1
urllib3==1.26.13
usb-creator==0.3.7
uvicorn==0.20.0
uvloop==0.17.0
wadllib==1.3.3
watchfiles==0.18.1
webdriver-manager==3.8.5
websockets==10.4
Werkzeug==0.16.1
wsproto==1.2.0
wxPython==4.0.7
xkit==0.0.0
